feat: Add support for downloading changelogs from changelogs.ubuntu.com
Using new arg `--use-changelogs-ubuntu-com` will download changelog direct from
changelogs.ubuntu.com instead of querying launchpad API.

This is useful if it is a kernel that was copied from a private PPA,
for example during an embargoed CVE where the public build for that source
package will not be the most recent and will not contain the most recent
changelog. This option assumes that the package is in the main Ubuntu archive.