feat: Add support for looking up changelog based on a binary package name
Example usage would be finding the source package changelog for a binary package install.

```
ubuntu-package-changelog --entries=1 focal Updates linux-aws-5.13-headers-5.13.0-1031
```

linux-aws-5.13-headers-5.13.0-1031 is built from source package linux-aws-5.13 but it is not easy to find this out.

Using this change to ubuntu-package-changelog you do not need to know and  ubuntu-package-changelog will find the
corresponding source package for you and get the changelog for that source package.